Using C# 7.3 so we can specify Enum use in generics
Added normal cast to integer types via dynamic
Split the PacketFactories, they now have too specific TIDType's.
Abstracted the Stream part of the serializer/deserializer
Fixed typos and formatting
Removed Vector types
Updated BouncyCastle
Updated version due to breaking changes

+++ src/Aragas.Network-2.Portable111/Data/VarZLong.cs
@@ -10,7 +10,7 @@ namespace Aragas.Network.Data
     [StructLayout(LayoutKind.Sequential, Pack = 1)]
     public struct VarZLong : IEquatable<VarZLong>
     {
-        public int Size => Variant.VariantSize((ulong) Variant.ZigZagEncode(_value));
+        public int Size => Variant.VariantSize(Variant.ZigZagEncode(_value));
 
 
         private readonly long _value;
